requirements.txt
# ===== Core =====
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ===== Data & Preprocessing =====
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# ===== NLP =====
transformers>=4.35.0
tokenizers>=0.15.0
datasets>=2.14.0

# ===== Multimodal =====
open_clip_torch>=2.20.0

# ===== Training & Experiment Tracking =====
tensorboard>=2.14.0
# wandb>=0.16.0

# ===== Visualization =====
matplotlib>=3.8.0
seaborn>=0.13.0

# ===== Notebooks =====
jupyter>=1.0.0
ipywidgets>=8.0.0

# ===== Augmentation =====
albumentations>=1.3.0
nlpaug>=1.1.0

# ===== Interpretability =====
captum>=0.7.0
grad-cam>=1.5.0

# ===== App Demo =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
# gradio>=4.0.0
streamlit>=1.28.0

# ===== Utilities =====
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
